Add some device creation/configuration.
Still somewhat a work in progress. Will die if we need to change type of device. Only
bridges and bonds are currently supported for creation. Unmanaged devices are still
checked and configured (if they have any files like 'master'). Making a device standalone
isn't yet supported. Fixing up configuration if devices are added/removed after startup
also doesn't happen. (e.g. if a regular device should be part of a bridge, we currently
only actually do that if the regular device exists at startup. If it is, e.g., USB
device and added later, we miss it.)