Misc messages.
Looks like we cannot add interfaces to masters we just created. Fetching
the master by name fails. If the user kills and restarts enetmgr, everything is
fine. So, probably the cache isn't being updated correctly.

Note that even waiting a while for messages doesn't work to get the updates
in the cache.